bumped to sc18

+++ src/util.cpp
@@ -27,7 +27,7 @@ bool is_double_str (const char * str) {
 }
 
 bool has_suffix (const char * str, const char * suffix) {
-  int k = strlen (str), l = strlen (suffix);
+  size_t k = strlen (str), l = strlen (suffix);
   return k > l && !strcmp (str + k - l, suffix);
 }
